lib/xe_ioctl: use valid type for bo
Fix bo handle type as it is uint32_t, not uint16_t.

Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>

+++ lib/xe/xe_ioctl.c
@@ -514,7 +514,7 @@ uint64_t xe_bo_mmap_offset(int fd, uint32_t bo)
 	return mmo.offset;
 }
 
-static void *__xe_bo_map(int fd, uint16_t bo, size_t size, int prot)
+static void *__xe_bo_map(int fd, uint32_t bo, size_t size, int prot)
 {
 	uint64_t mmo;
 	void *map;
